+++ OMS.API/OMS.Domain.Entities/API/Response/Organization/GetOrganizationLogisticDetailsResponse.cs
@@ -8,7 +8,7 @@ namespace OMS.Domain.Entities.API.Response.Organization
 {
     public class GetOrganizationLogisticDetailsResponse
     {
-        public short? OrganizationLogisticDetailId {  get; set; }
+        public byte? OrganizationLogisticDetailId {  get; set; }
         public string? FedExAccount {  get; set; }
         public string? DHLAccount { get; set; }
         public string? UPSAccount {  get; set; }
