+++ OMS.API/OMS.Domain.Entities/API/Response/Organization/GetOrganizationShippingChargesResponse.cs
@@ -8,7 +8,7 @@ namespace OMS.Domain.Entities.API.Response.Organization
 {
     public class GetOrganizationShippingChargesResponse
     {
-        public short? OrganizationShippingChargeId { get; set; }
+        public byte? OrganizationShippingChargeId { get; set; }
         public decimal? DomesticOvernight { get; set; }
         public decimal? DomesticSecondDay { get; set; }
         public decimal? DomesticGround { get; set; }
