+++ OMS.API/OMS.Domain.Entities/Entity/Organization/OrganizationAccountingDetailsDto.cs
@@ -8,7 +8,7 @@ namespace OMS.Domain.Entities.Entity.Organization
 {
     public class OrganizationAccountingDetailsDto
     {
-        public short? OrganizationAccountingDetailId { get; set; }
+        public byte? OrganizationAccountingDetailId { get; set; }
         public decimal? CreditLimit { get; set; }
         public DateTime? CreatedAt { get; set; }
         public short? CreatedBy { get; set; }
